moved KnitProgress classes to ayab_knitprogress.py
