+++ src/main/python/ayab/plugins/ayab_plugin/ayab_status.py
@@ -17,11 +17,8 @@
 #    Copyright 2014 Sebastian Oliva, Christian Obersteiner, Andreas Müller, Christian Gerbrandt
 #    https://github.com/AllYarnsAreBeautiful/ayab-desktop
 
-from PyQt5 import QtWidgets, QtCore
-from PyQt5.QtCore import Qt, QCoreApplication
-from PyQt5.QtWidgets import QGridLayout, QLayout, QLabel
+from PyQt5.QtCore import QCoreApplication
 from bitarray import bitarray
-import numpy as np
 
 
 class Status(object):
@@ -68,98 +65,3 @@ class Status(object):
         self.hall_r = hall_r
         self.carriage_type = carriage_type
         self.carriage_position = carriage_position
-
-
-class KnitProgress(object):
-    """Methods for knit progress window."""
-    def __init__(self, parent):
-        self.area = parent.area
-        self.area.setContentsMargins(1, 1, 1, 1)
-
-    def reset(self):
-        self.container = QtWidgets.QWidget()
-        self.container.setMinimumSize(100, 100)
-        self.grid = QGridLayout(self.container)
-        self.grid.setContentsMargins(1, 1, 1, 1)
-        self.grid.setSpacing(0)
-        self.grid.setSizeConstraint(QLayout.SetMinAndMaxSize)
-        self.area.setWidget(self.container)
-        self.row = -1
-
-    def update(self, status, row_multiplier):
-        if status.current_row < 0:
-            return
-        _translate = QtCore.QCoreApplication.translate
-        row, swipe = divmod(status.line_number, row_multiplier)
-        direction = status.line_number % 2
-        # row
-        w0 = self.__label(_translate("KnitProgress", "Row"))
-        self.grid.addWidget(w0, status.line_number, 0)
-        w1 = self.__label(str(status.current_row))
-        self.grid.addWidget(w1, status.line_number, 1, 1, 1, Qt.AlignRight)
-        # pass
-        w2 = self.__label(
-            _translate("KnitProgress", "Pass") + " " + str(swipe + 1))
-        self.grid.addWidget(w2, status.line_number, 2)
-        # color
-        if status.color_symbol == "":
-            coltext = ""
-        else:
-            coltext = _translate("KnitProgress",
-                                 "Color") + " " + status.color_symbol
-        w3 = self.__label(coltext)
-        self.grid.addWidget(w3, status.line_number, 3)
-        # carriage and direction
-        try:
-            carriage = status.carriage[0] + status.carriage[2] + " "
-        except Exception:
-            carriage = ""
-        w4 = self.__label(carriage + ["\u2192 ", "\u2190 "][direction])
-        self.grid.addWidget(w4, status.line_number, 4)
-        # TODO: hints, notes, memos
-        w0.show()
-        w1.show()
-        w2.show()
-        w3.show()
-        w4.show()
-        self.area.ensureWidgetVisible(w0)
-        # graph line of stitches
-        for c in range(len(status.bits)):
-            wc = self.__stitch(status.color, status.bits[c], status.alt_color)
-            self.grid.addWidget(wc, status.line_number, 6 + c)
-        return
-
-    def __label(self, text):
-        table = "<table><tbody><tr height='12'><td style='font-weight: normal;'>" + text + "</td></tr></tbody></table>"
-        label = QLabel()
-        label.setText(table)
-        label.setTextFormat(Qt.RichText)
-        return label
-
-    def __stitch(self, color, bit, alt_color=None, symbol=0x20):
-        table = "<table style='font-weight: normal;'><tbody><tr height='12'><td width='12' align='center' "
-        if bit:
-            table = table + "style='border: 1 solid black; color: #{:06x}; background-color: #{:06x};'>".format(
-                self.__contrast_color(color), color) + chr(symbol)
-        elif alt_color is not None:
-            table = table + "style='border: 1 solid black; color: #{:06x}; background-color: #{:06x};'>".format(
-                self.__contrast_color(alt_color), alt_color) + chr(symbol)
-        else:
-            table = table + "style='border: 1 dotted black;'>"
-        table = table + "</td></tr></tbody></table>"
-        label = QLabel()
-        label.setText(table)
-        label.setTextFormat(Qt.RichText)
-        return label
-
-    # functions used to calculate contrasting text color
-    def __rgb2array(self, color):
-        return np.array(
-            [color // 0x1000, (color // 0x100) & 0xFF, color & 0xFF])
-
-    def __greyscale(self, rgb):
-        return np.dot([.299, .587, .114], rgb)
-
-    def __contrast_color(self, color):
-        return [0x000000,
-                0xFFFFFF][self.__greyscale(self.__rgb2array(color)) < 0x80]
